+++ src/main/java/net/chaos/chaosmod/world/events/PlayerTickBiomeEvent.java
@@ -4,6 +4,7 @@ import java.util.Set;
 
 import net.chaos.chaosmod.common.capabilities.biome.CapabilityVisitedBiomes;
 import net.chaos.chaosmod.common.capabilities.biome.VisitedBiomes;
+import net.chaos.chaosmod.jobs.events.traveler.JobTravelerEventHandler;
 import net.chaos.chaosmod.network.packets.PacketManager;
 import net.minecraft.entity.player.EntityPlayer;
 import net.minecraft.entity.player.EntityPlayerMP;
@@ -39,13 +40,16 @@ public class PlayerTickBiomeEvent {
 		if (visitedBiomes.contains(currentBiome))
 			return;
 
+		// FIXME : not working
+		if (!JobTravelerEventHandler.onPlayerDiscoversBiome(player, currentBiome)) return;
+
 		String data = currentBiome.getRegistryName().toString();
 
 		PacketManager.network.sendTo(new MessageDisplayBiomeName(data), (EntityPlayerMP) player);
 
 		world.playSound(null, player.getPosition(), SoundEvents.ENTITY_EXPERIENCE_ORB_PICKUP, SoundCategory.PLAYERS,
 				1.0f, 1.0f);
-
+		
 		biomes.addBiome(currentBiome);
 	}
 }